firmware: sync latest characterization/validation work
Mirror the current working tree: fp validation benchmark ffma (fp_fma),
idle-baseline reset fix in verify.py, GUI batch cap raise + scatter sizing,
plus the accumulated loops/regression kernels and scripts.

+++ firmware/regression/sources/wl_ratscale.c
@@ -5,7 +5,7 @@
 static int A[N]; static int init=0;
 void run_workload(void){
   volatile int sink=0;
-  if(!init){ for(int i=0;i<N;i++) A[i]=(i*1103515245+12345)|1; init=1; }
+  if(!init){ for(int i=0;i<N;i++) A[i]=(int)(((unsigned)i*1103515245u+12345u)|1u); init=1; }
   for(int r=0;r<REPS;r++){
     int d=(r&255)+3;
     for(int i=0;i<N;i++){
